requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
flake8==4.0.1
flake8-bugbear==21.9.2
flake8-polyfill==1.0.2
idna==3.3
importlib-metadata==4.2.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.2
pep8-naming==0.12.1
Pillow==8.3.2
prompt-toolkit==3.0.22
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
